Basic Installation JDK ,GITHUB
First, we downloaded and installed the JDK (Java Development Kit), which is essential for compiling and running Java programs on your computer. After that, we created a GitHub account to host and manage our code repositories, enabling version control and collaboration. Finally, we set up IntelliJ IDEA, a powerful IDE, and connected it to GitHub, allowing us to directly push and pull code between the IDE and our GitHub repository.